+++ src/main/java/gift/entity/Product.java
@@ -1,4 +1,4 @@
-package gift.model;
+package gift.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
