+++ src/main/java/gift/util/UserUtility.java
@@ -1,6 +1,6 @@
 package gift.util;
 
-import gift.model.User;
+import gift.entity.User;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
 import io.jsonwebtoken.JwtException;
