+++ src/main/java/gift/validation/constraint/NameConstraint.java
@@ -1,16 +1,17 @@
-package gift.util;
+package gift.validation.constraint;
 
+import gift.validation.validator.ProductNameValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
 import java.lang.annotation.*;
 
 @Documented
-@Constraint(validatedBy = ProductIdValidator.class)
+@Constraint(validatedBy = ProductNameValidator.class)
 @Target({ElementType.METHOD, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface ProductIdConstraint {
-    String message() default "Invalid product";
+public @interface NameConstraint {
+    String message() default "Invalid name";
 
     Class<?>[] groups() default {};
