+++ src/main/java/gift/validation/constraint/ProductIdConstraint.java
@@ -1,16 +1,17 @@
-package gift.util;
+package gift.validation.constraint;
 
+import gift.validation.validator.ProductIdValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
 import java.lang.annotation.*;
 
 @Documented
-@Constraint(validatedBy = ProductNameValidator.class)
+@Constraint(validatedBy = ProductIdValidator.class)
 @Target({ElementType.METHOD, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface NameConstraint {
-    String message() default "Invalid name";
+public @interface ProductIdConstraint {
+    String message() default "Invalid product";
 
     Class<?>[] groups() default {};
